feat: :fire: access based control guards in controller methods, ui menus, table dropdown and more

+++ app/Http/Controllers/PatientController.php
@@ -14,18 +14,34 @@ class PatientController extends Controller
      */
     public function index(Request $request)
     {
+        // View Access
+        if ($request->user()->cannot('view_patients')) {
+            abort(403);
+        }
+
         $patients = Patient::query()->when($request->input('search'), function ($query, $search) {
             $query->where('name', 'like', "%{$search}%");
         })->orderBy('created_at', 'DESC')->paginate()->withQueryString();
 
-        return inertia('Dashboard/Patients/Index', ['patients' => $patients, 'filters' => $request->only('search')]);
+        return inertia('Dashboard/Patients/Index', ['patients' => $patients, 'filters' => $request->only('search'), 'can' => [
+            'view_patients' => $request->user()->can('view_patients'),
+            'show_patient' => $request->user()->can('show_patient'),
+            'create_patient' => $request->user()->can('create_patient'),
+            'edit_patient' => $request->user()->can('edit_patient'),
+            'delete_patient' => $request->user()->can('delete_patient')
+        ]]);
     }
 
     /**
      * Show the form for creating a new resource.
      */
-    public function create()
+    public function create(Request $request)
     {
+        // Create Access
+        if ($request->user()->cannot('create_patient')) {
+            abort(403);
+        }
+
         return inertia('Dashboard/Patients/Add');
     }
 
@@ -34,6 +50,11 @@ class PatientController extends Controller
      */
     public function store(StorePatientRequest $request)
     {
+        // Create Access
+        if ($request->user()->cannot('create_patient')) {
+            abort(403);
+        }
+
         Patient::create($request->validated());
 
         return redirect(route('patients.index'));
@@ -42,8 +63,13 @@ class PatientController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(Patient $patient)
+    public function show(Request $request, Patient $patient)
     {
+        // Show Access
+        if ($request->user()->cannot('show_patient')) {
+            abort(403);
+        }
+
         $patient = Patient::findOrFail($patient->id);
         $records = $patient->records()->latest()->paginate();
         return inertia('Dashboard/Patients/Show', ['patient' => $patient, 'records' => $records]);
@@ -52,8 +78,13 @@ class PatientController extends Controller
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(Patient $patient)
+    public function edit(Request $request, Patient $patient)
     {
+        // Edit Access
+        if ($request->user()->cannot('edit_patient')) {
+            abort(403);
+        }
+
         $patient = Patient::findOrFail($patient->id);
         return inertia('Dashboard/Patients/Edit', ['patient' => $patient]);
     }
@@ -63,6 +94,11 @@ class PatientController extends Controller
      */
     public function update(UpdatePatientRequest $request, Patient $patient)
     {
+        // Edit Access
+        if ($request->user()->cannot('edit_patient')) {
+            abort(403);
+        }
+
         $patient = Patient::findOrFail($patient->id);
         $patient->update($request->validated());
 
@@ -72,8 +108,13 @@ class PatientController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Patient $patient)
+    public function destroy(Request $request, Patient $patient)
     {
+        // Delete Access
+        if ($request->user()->cannot('delete_patient')) {
+            abort(403);
+        }
+
         $patient->delete();
 
         return redirect()->back();
